Move EventSourceState to separate file
