+++ src/WinRT.Runtime/Interop/EventSource.cs
@@ -10,116 +10,6 @@ using System.Threading;
 
 namespace WinRT.Interop
 {
-    // Registration state and delegate cached separately to survive EventSource garbage collection
-    // and to prevent the generated event delegate from impacting the lifetime of the
-    // event source.
-    internal abstract class EventSourceState<TDelegate> : IDisposable
-        where TDelegate : class, MulticastDelegate
-    {
-        public EventRegistrationToken token;
-        public System.Delegate del;
-        public System.Delegate eventInvoke;
-        private bool disposedValue;
-        private readonly IntPtr obj;
-        private readonly int index;
-        private readonly System.WeakReference<object> cacheEntry;
-        private IntPtr eventInvokePtr;
-        private IntPtr referenceTrackerTargetPtr;
-
-        protected EventSourceState(IntPtr obj, int index)
-        {
-            this.obj = obj;
-            this.index = index;
-            eventInvoke = GetEventInvoke();
-            cacheEntry = new System.WeakReference<object>(this);
-        }
-
-        // The lifetime of this object is managed by the delegate / eventInvoke
-        // through its target reference to it.  Once the delegate no longer has
-        // any references, this object will also no longer have any references.
-        ~EventSourceState()
-        {
-            Dispose(false);
-        }
-
-        // Allows to retrieve a singleton like weak reference to use
-        // with the cache to allow for proper removal with comparision.
-        public System.WeakReference<object> GetWeakReferenceForCache()
-        {
-            return cacheEntry;
-        }
-
-        protected abstract TDelegate GetEventInvoke();
-
-        protected virtual void Dispose(bool disposing)
-        {
-            // Uses the dispose pattern to ensure we only remove
-            // from the cache once: either via unsubscribe or via
-            // the finalizer.
-            if (!disposedValue)
-            {
-                EventSourceCache.Remove(obj, index, cacheEntry);
-                disposedValue = true;
-            }
-        }
-
-        public void Dispose()
-        {
-            Dispose(true);
-            GC.SuppressFinalize(this);
-        }
-
-        public void InitalizeReferenceTracking(IntPtr ptr)
-        {
-            eventInvokePtr = ptr;
-            int hr = Marshal.QueryInterface(ptr, ref Unsafe.AsRef(in IID.IID_IReferenceTrackerTarget), out referenceTrackerTargetPtr);
-            if (hr != 0)
-            {
-                referenceTrackerTargetPtr = default;
-            }
-            else
-            {
-                // We don't want to keep ourselves alive and as long as this object
-                // is alive, the CCW still exists.
-                Marshal.Release(referenceTrackerTargetPtr);
-            }
-        }
-
-        public unsafe bool HasComReferences()
-        {
-            if (eventInvokePtr != default)
-            {
-                IUnknownVftbl vftblIUnknown = **(IUnknownVftbl**)eventInvokePtr;
-                vftblIUnknown.AddRef(eventInvokePtr);
-                uint comRefCount = vftblIUnknown.Release(eventInvokePtr);
-                if (comRefCount != 0)
-                {
-                    return true;
-                }
-            }
-
-            if (referenceTrackerTargetPtr != default)
-            {
-                void** vftblReferenceTracker = *(void***)referenceTrackerTargetPtr;
-
-                // AddRefFromReferenceTracker
-                _ = ((delegate* unmanaged[Stdcall]<IntPtr, uint>)(vftblReferenceTracker[3]))(referenceTrackerTargetPtr);
-
-                // ReleaseFromReferenceTracker
-                uint refTrackerCount = ((delegate* unmanaged[Stdcall]<IntPtr, uint>)(vftblReferenceTracker[4]))(referenceTrackerTargetPtr);
-
-                if (refTrackerCount != 0)
-                {
-                    // Note we can't tell if the reference tracker ref is pegged or not, so this is best effort where if there
-                    // are any reference tracker references, we assume the event has references.
-                    return true;
-                }
-            }
-
-            return false;
-        }
-    }
-
     internal unsafe abstract class EventSource<TDelegate>
         where TDelegate : class, MulticastDelegate
     {
